% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibagInputModule.R
\name{ibagInputUi}
\alias{ibagInputUi}
\title{UI module}
\usage{
ibagInputUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module
}
