% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibagIntroModule.R
\name{ibagIntroUi}
\alias{ibagIntroUi}
\title{Intro UI module}
\usage{
ibagIntroUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Intro UI module
}
