% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibagLinearModule.R
\name{ibagLinear}
\alias{ibagLinear}
\title{Server module generating the pokemon interface}
\usage{
ibagLinear(input, output, session)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}
}
\description{
Server module generating the pokemon interface
}
