% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibagLinearModule.R
\name{ibagLinearUi}
\alias{ibagLinearUi}
\title{Linear module}
\usage{
ibagLinearUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Linear module
}
